refactor: split backend/service.py (2522 lines) into service/ package
Decomposed monolithic service.py into 6 focused modules:
- core.py: CorridorKeyService class, dataclasses, properties
- model_manager.py: VRAM residency, model loading/unloading
- inference.py: sequential + parallel inference pipelines
- pipelines.py: GVM, SAM2, VideoMaMa, MatAnyone2, BiRefNet
- frame_ops.py: frame I/O, mask reading, output writing
- helpers.py: standalone utilities, MatAnyone2 import logic

Public API unchanged — all existing imports still work.
Threading contracts (gpu_lock, inference gating) preserved.
